- New progress system: dynamic required progress (uses expressions now) - Performance & stability improvements for string replacements in languagemanager - Add ActiveObjective as possible argument for objective task descriptions (breaks API)

+++ paper/src/main/java/rocks/gravili/notquests/paper/events/hooks/CitizensEvents.java
@@ -124,7 +124,7 @@ public class CitizensEvents implements Listener {
                                                     continue;
                                                 }
 
-                                                final long progressLeft = activeObjective.getProgressNeeded() - activeObjective.getCurrentProgress();
+                                                final double progressLeft = activeObjective.getProgressNeeded() - activeObjective.getCurrentProgress();
 
                                                 if (progressLeft == 0) {
                                                     continue;
